Split mempool into tx and da mempool
