+++ nomos-services/api/src/http/mempool.rs
@@ -1,12 +1,12 @@
 use core::{fmt::Debug, hash::Hash};
 use nomos_core::header::HeaderId;
 use nomos_mempool::{
-    backend::mockpool::MockPool, network::NetworkAdapter, Discriminant, MempoolMsg, MempoolService,
+    backend::mockpool::MockPool, network::NetworkAdapter, TxMempoolMsg, TxMempoolService,
 };
 use nomos_network::backends::NetworkBackend;
 use tokio::sync::oneshot;
 
-pub async fn add<N, A, D, Item, Key>(
+pub async fn add<N, A, Item, Key>(
     handle: &overwatch_rs::overwatch::handle::OverwatchHandle,
     item: Item,
     converter: impl Fn(&Item) -> Key,
@@ -15,18 +15,17 @@ where
     N: NetworkBackend,
     A: NetworkAdapter<Backend = N, Item = Item, Key = Key> + Send + Sync + 'static,
     A::Settings: Send + Sync,
-    D: Discriminant,
     Item: Clone + Debug + Send + Sync + 'static + Hash,
     Key: Clone + Debug + Ord + Hash + 'static,
 {
     let relay = handle
-        .relay::<MempoolService<A, MockPool<HeaderId, Item, Key>, D>>()
+        .relay::<TxMempoolService<A, MockPool<HeaderId, Item, Key>>>()
         .connect()
         .await?;
     let (sender, receiver) = oneshot::channel();
 
     relay
-        .send(MempoolMsg::Add {
+        .send(TxMempoolMsg::Add {
             key: converter(&item),
             item,
             reply_channel: sender,
