+++ nomos-services/mempool/src/da/mod.rs
@@ -1,0 +1,1 @@
+pub mod service;
