+++ nomos-services/mempool/src/network/messages.rs
@@ -4,6 +4,6 @@ use serde::{Deserialize, Serialize};
 // internal
 
 #[derive(Serialize, Deserialize)]
-pub struct TransactionMsg<Tx> {
-    pub tx: Tx,
+pub struct PayloadMsg<Payload> {
+    pub payload: Payload,
 }
