+++ nomos-services/mempool/src/tx/mod.rs
@@ -1,0 +1,3 @@
+#[cfg(feature = "metrics")]
+pub mod metrics;
+pub mod service;
